refactor(blorktools): css3d scene helper
